Handling Exception

+++ CriminalEvidenceManagementSystem/src/main/java/com/fis/cems/dao/mem/CriminalCaseMemDAO.java
@@ -1,10 +1,14 @@
 package com.fis.cems.dao.mem;
 
 import com.fis.cems.dao.ICriminalCaseDAO;
+import com.fis.cems.exception.CriminalCaseNotFoundException;
+import com.fis.cems.exception.InvalidParamException;
 import com.fis.cems.model.CriminalCase;
+import com.fis.cems.model.enums.CaseStatus;
 import com.fis.cems.model.enums.CaseType;
 import com.fis.cems.util.MemoryStorageUtil;
 
+import javax.management.InvalidAttributeValueException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -15,7 +19,7 @@ public class CriminalCaseMemDAO implements ICriminalCaseDAO {
     private final static Map<Long, CriminalCase> database = MemoryStorageUtil.getCriminalCases();
     @Override
     public List<CriminalCase> findAll() {
-        return new ArrayList<>(database.values());
+        return database.values().stream().toList();
     }
 
     @Override
@@ -27,15 +31,20 @@ public class CriminalCaseMemDAO implements ICriminalCaseDAO {
 
     @Override
     public CriminalCase save(CriminalCase criminalCase) {
-        if (criminalCase.getId() < 0) return null;
 
-        if (criminalCase.getId() == null) {
+        if (criminalCase.getId() == null)
             criminalCase.setId(database.size() + 1L);
-        }
+
+        if (criminalCase.getId() <= 0)
+            throw new InvalidParamException("Invalid criminal case with id: " + criminalCase.getId());
+
+        if (!database.containsKey(criminalCase.getId()))
+            criminalCase.setId(database.size() + 1L);
 
         database.put(criminalCase.getId(), criminalCase);
 
         return database.get(criminalCase.getId());
+
     }
 
     @Override
@@ -45,6 +54,12 @@ public class CriminalCaseMemDAO implements ICriminalCaseDAO {
 
     @Override
     public void delete(CriminalCase criminalCase) {
+        if (criminalCase.getId() == null || criminalCase.getId() <= 0)
+            throw new InvalidParamException("Invalid criminal case with id: " + criminalCase.getId());
+
+        if (findById(criminalCase.getId()).isEmpty())
+            throw new CriminalCaseNotFoundException("CriminalCase with id: " + criminalCase.getId() + " not found!");
+
         database.remove(criminalCase.getId(), criminalCase);
     }
 
@@ -54,7 +69,11 @@ public class CriminalCaseMemDAO implements ICriminalCaseDAO {
     }
 
     @Override
-    public List<CriminalCase> findByCaseType(CaseType caseType) {
-        return null;
+    public List<CriminalCase> findByCaseStatusOnlyForChiefRank(CaseStatus caseStatus) {
+        return database.values()
+                .stream()
+                .filter(criminalCase -> criminalCase.getStatus() == caseStatus)
+                .collect(Collectors.toList());
     }
+
 }
